Update pcboot to work with newer Rust and sfdisk versions
rustc 1.15.0-nightly (3bf2be9ce 2016-11-22)

According to the manpage, sfdisk 2.26 removed the CHS parameters, so stop
specifying them in mk/test.sh.  (Yes, obviously disks don't actually have
a true CHS geometry, but those fields still exist in the partition table.
Is sfdisk making them up now?  Of course, it works anyway.)

+++ src/stage1/panic.rs
@@ -13,12 +13,12 @@ impl fmt::Write for SimpleWriter {
     }
 }
 
-#[lang = "panic_fmt"] #[cold] #[inline]
-extern fn rust_panic_fmt(msg: fmt::Arguments, file: &'static str, line: usize) -> ! {
+#[lang = "panic_fmt"] #[cold]
+extern fn rust_panic_fmt(msg: fmt::Arguments, file: &'static str, line: u32) -> ! {
     sys::print_str(&"internal error: ");
     sys::print_str(&file);
     sys::print_str(&":");
-    sys::print_u32(line as u32);
+    sys::print_u32(line);
     sys::print_str(&": ");
     let _ = fmt::write(&mut SimpleWriter, msg);
     sys::halt();
